+++ test/testMNIST/testMNIST.cpp
@@ -63,7 +63,7 @@ int main(int argc, const char** args){
     args_vec.push_back("-norm");
 
 
-    long opti = sol_init_optimizer(dataset, loss_func,args_vec.size(), &(args_vec[0]));
+    long opti = sol_init_optimizer(dataset, loss_func,(int)(args_vec.size()), &(args_vec[0]));
     if  (opti == 0){
         delete reader;
         release(dataset, loss_func, opti);
